Moving geoapify key into env vars. Fixing websockets on docker
