+++ .github/workflows/functional_testing.yml
@@ -29,6 +29,9 @@ jobs:
           echo "# api information" >> .env
           echo "API_PORT=3001" >> .env
           echo "API_LIMIT=10000" >> .env
+          echo "" >> .env
+          echo "# third party information" >> .env
+          echo "GEOAPIFY_API_KEY=${{ secrets.GEOAPIFY_API_KEY }}" >> .env
 
       - name: Build Dockerfile 🐋
         run: docker-compose -f docker-compose.yml up -d --build
