+++ docker-compose.yml
@@ -42,6 +42,10 @@ services:
     build:
       context: ui
       dockerfile: Dockerfile
+      args:
+        REACT_APP_API_PORT: ${API_PORT}
+        REACT_APP_HTTP_PORT: ${HTTP_PORT}
+        REACT_APP_GEOAPIFY_API_KEY: ${GEOAPIFY_API_KEY}
     container_name: snnap_ui
     restart: always
     depends_on:
@@ -50,8 +54,6 @@ services:
       - ${HTTP_PORT}:80
     links:
       - api
-    environment:
-      API_PORT: ${API_PORT}
-      GEOAPIFY_API_KEY: ${GEOAPIFY_API_KEY}
+
 volumes:
   mysql:
