+++ ui/nginx.conf
@@ -21,4 +21,12 @@ server {
   location /api {
     proxy_pass http://api;
   }
+
+  location /wsapp {
+    proxy_pass http://api;
+    proxy_http_version 1.1;
+    proxy_set_header Upgrade $http_upgrade;
+    proxy_set_header Connection "Upgrade";
+    proxy_set_header Host $host;
+  }
 }
